moved libraris from domens into auth
