+++ libs/auth/feature-register-client/.babelrc
