+++ libs/auth/feature-register-client/eslint.config.mjs
