+++ libs/auth/feature-register-client/package.json
