+++ libs/auth/feature-register-client/tsconfig.json
