+++ libs/auth/feature-register-client/tsconfig.lib.json
