+++ libs/auth/feature-register-client/tsconfig.spec.json
