+++ libs/auth/feature-register-client/vite.config.ts
@@ -6,7 +6,7 @@ import * as path from 'path';
 
 export default defineConfig(() => ({
   root: __dirname,
-  cacheDir: '../../../node_modules/.vite/libs/specialist/feature-register',
+  cacheDir: '../../../node_modules/.vite/libs/auth/feature-register-client',
   plugins: [
     react(),
     dts({
@@ -30,7 +30,7 @@ export default defineConfig(() => ({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: 'src/index.ts',
-      name: '@paritet/feature-register-specialist',
+      name: '@paritet/feature-register-client',
       fileName: 'index',
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
