+++ libs/auth/feature-register-specialist/package.json
