+++ libs/auth/feature-register-specialist/src/index.ts
