+++ libs/auth/feature-register-specialist/src/lib/components/SpecialistRegistrationForm.spec.tsx
