+++ libs/auth/feature-register-specialist/tsconfig.json
