+++ libs/auth/feature-register-specialist/tsconfig.lib.json
