+++ libs/auth/feature-register-specialist/tsconfig.spec.json
