+++ libs/auth/feature-register-specialist/vite.config.ts
@@ -6,7 +6,7 @@ import * as path from 'path';
 
 export default defineConfig(() => ({
   root: __dirname,
-  cacheDir: '../../../node_modules/.vite/libs/client/feature-register',
+  cacheDir: '../../../node_modules/.vite/libs/auth/feature-register-specialist',
   plugins: [
     react(),
     dts({
@@ -30,7 +30,7 @@ export default defineConfig(() => ({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: 'src/index.ts',
-      name: '@paritet/feature-register-client',
+      name: '@paritet/feature-register-specialist',
       fileName: 'index',
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
